Allow type conversion from BufferStatus to int
To get around the compiler error
error: format specifies type int but the argument has type BufferStatus

Bug: b/312565119

+++ media/module/bufferpool/2.0/AccessorImpl.cpp
@@ -609,7 +609,7 @@ void Accessor::Impl::BufferPool::processStatusMessages() {
         }
         if (ret == false) {
             ALOGW("buffer status message processing failure - message : %d connection : %lld",
-                  message.newStatus, (long long)message.connectionId);
+                  (int)message.newStatus, (long long)message.connectionId);
         }
     }
     messages.clear();
